+++ lib/shipcloud/base.rb
@@ -26,5 +26,13 @@ module Shipcloud
     def self.base_url
       "#{class_name.downcase}s"
     end
+
+    def self.camel_to_snakecase(string)
+      string.gsub(/::/, "/").
+        gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2').
+        gsub(/([a-z\d])([A-Z])/, '\1_\2').
+        tr("-", "_").
+        downcase
+    end
   end
 end